SecondLab/src/main/java/by/bsuir/task2/service/validation/Validator.java
package by.bsuir.task02.service.validation;

import by.bsuir.task02.entity.criteria.Criteria;

public class Validator {

    public static boolean isCriteriaValid(Criteria criteria) {
        if (criteria == null) {
            return false;
        }

        return criteria.getCriteriaMap().entrySet().stream().allMatch(
                entry -> CriteriaValidatorFactory.getInstance()
                        .getValidator(entry.getKey())
                        .isCriteriaValid(entry.getValue()));
    }

}